drop unsigned longs in favor of size_t

+++ src/include/buffer.h
@@ -1,9 +1,11 @@
 #ifndef BUFFER_H
 #define BUFFER_H
 
+#include <stddef.h>
+
 typedef struct {
     char *data;
-    unsigned long length;
+    size_t length;
 } Buffer;
 
 Buffer *buffer_init();
