+++ src/include/parser.h
@@ -8,8 +8,8 @@
 typedef struct {
     Token **tokens;
     Token *current;
-    unsigned long position;
-    unsigned long length;
+    size_t position;
+    size_t length;
 } Parser;
 
 Parser *parser_init(Token **tokens);
